Fixing datatype for box size: usize -> u64
The arbitrary size limit in heif/boxes/meta.rs of usize::MAX - 1000 made me look up if the specifications have anything more specific to say about this. While taking a closer look at this, I noticed somewhat of a design mistake, namely that usize may not be the best option for describing box sizes, as boxes may be larger than u32::MAX bytes but 32-bit systems can't express that using usize.

+++ src/heif/boxes/iso.rs
@@ -60,7 +60,7 @@ IsoBox
         let mut buffer: Vec<u8> = Vec::new();
 
         // This may cause an out of memory error, but won't panic like vec![]
-        buffer.try_reserve_exact(data_left_to_read)?;
+        buffer.try_reserve_exact(data_left_to_read as usize)?;
 
         // Can't use read_exact here because the name buffer we read into is
         // still size 0 (only has reserved capacity!)
